manage fields not immediately present in DOM
There are some cases where fields may not immediately be present in the
DOM and added later (GitHub Slash Commands for instance). In these
cases, when the fields are restored, those not present in the DOM are
not restored (because they simply don't exist yet), however we completely
remove all the stored fields' values from the session store. This means that
fields that are added later will not have a change to restore their data.

This change holds onto field values for fields that weren't able to be
located in case that field is added to the DOM later and can then be
restored.

If a textarea never appears, it would stay in the session storage until the
tab is closed. Because session storage is scoped to the tab, automatically
cleared when the tab is closed, and has a 5 MB limit, this is unlikely to cause
memory bloat problems.