Make sure to access `sessionStorage` in a `try` block
In https://github.com/github/session-resume/pull/38 we allowed callers
to pass in their own storage implementation. We accidentally leaked a `window.sessionStorage` access outside of a `try` block, though,
and this could lead to errors when users have cookies disabled or are
in private mode.

I've chosen to keep the other `try/catch` blocks to minimize changes.